Fixed #3

+++ core/components/cmpgenerator/processors/mgr/cmpgenerator/build.php
@@ -75,23 +75,27 @@ if ($cmp->save() == false) {
 
 // 1 delete all files for testing:
 function rmdir_files($dir) {
-    foreach( glob( $dir . '*', GLOB_MARK ) as $file) {
-        if (is_dir($file)) {
-            rmdir_files($file."/");
-            if( is_dir($file) ) {
-                rmdir($file);
+    $glob = glob( $dir . '*', GLOB_MARK );
+    if (is_array($glob)) {
+        foreach ($glob as $file) {
+            if (is_dir($file)) {
+                rmdir_files($file."/");
+                if( is_dir($file) ) {
+                    rmdir($file);
+                }
+            } else if( is_file($file) )  {
+                unlink($file);
             }
-        } elseif( is_file($file) )  {
-            unlink($file);
         }
     }
     if (is_dir($dir) ){
-        if( rmdir( $dir ) ){
+        if ( rmdir( $dir ) ){
             return true;
         }
         return false;
     }
-}
+} 
+ 
 $directories = array(
     'assets' => MODX_ASSETS_PATH.'components/'.$package_name.'/',
     'code_base' => MODX_CORE_PATH.'components/'.$package_name.'/'
